+++ R/manipulation.R
@@ -96,10 +96,15 @@ add_row_cols <- function (x, y, after, dimno, ...) {
 #' @param after Insert the row/column after this position. 0 (the default) inserts as the first row/column.
 #' @param fill String. If `...` contains fewer elements than there are columns/rows to
 #' fill, the remaining cells will be filled with this.
+#' @param rowspan,colspan Scalar integer. Sets the rowspan or colspan of the *first* cell only.
 #' this. The default `NULL` throws an error if there are too few elements.
 #' @param copy_cell_props Copy cell properties from the previous row or column (if after > 0). See [cbind.huxtable()].
 #' @details
 #' In `insert_column` only, you can use a column name for `after`.
+#'
+#' Even if `colspan` or `rowspan` are greater than 1, you must still provide
+#' values for the hidden cells. Use `fill = ""` for this.
+#'
 #' @return The modified huxtable
 #' @seealso [add_rows()] and [add_columns()], which insert multiple rows/columns at once.
 #' @export
@@ -120,12 +125,21 @@ add_row_cols <- function (x, y, after, dimno, ...) {
 #'         after = "Price",
 #'         fill = "50%"
 #'       )
-insert_column <- function (ht, ..., after = 0, fill = NULL, copy_cell_props = TRUE) {
+#'
+#' # don't forget to use `fill`:
+#' insert_row(jams,
+#'         "Jams and prices",
+#'         fill = "",
+#'         colspan = 2
+#'       )
+insert_column <- function (ht, ..., after = 0, fill = NULL, rowspan = 1,
+      copy_cell_props = TRUE) {
   # is.count would complain about 0
   assert_that(
           is.scalar(after),
           is.number(after) || is.string(after),
-          is.null(fill) || is.scalar(fill)
+          is.null(fill) || is.scalar(fill),
+          is.count(rowspan)
         )
   if (is.number(after)) assert_that(after >= 0, after <= ncol(ht))
   if (is.string(after)) {
@@ -148,6 +162,8 @@ insert_column <- function (ht, ..., after = 0, fill = NULL, copy_cell_props = TR
   res <- if (! is.null(ht1)) cbind(ht1, to_insert, copy_cell_props = copy_cell_props) else to_insert
   res <- if (! is.null(ht2)) cbind(res, ht2) else res
 
+  rowspan(res)[1, after + 1] <- rowspan
+
   res
 }
 
@@ -155,13 +171,14 @@ insert_column <- function (ht, ..., after = 0, fill = NULL, copy_cell_props = TR
 #' @rdname insert_column
 #'
 #' @export
-insert_row <- function (ht, ..., after = 0, fill = NULL, copy_cell_props = TRUE) {
+insert_row <- function (ht, ..., after = 0, fill = NULL, colspan = 1, copy_cell_props = TRUE) {
   assert_that(
           is.scalar(after),
           is.number(after),
           after >= 0,
           after <= nrow(ht),
-          is.null(fill) || is.scalar(fill)
+          is.null(fill) || is.scalar(fill),
+          is.count(colspan)
         )
 
   ht1 <- NULL
@@ -178,6 +195,8 @@ insert_row <- function (ht, ..., after = 0, fill = NULL, copy_cell_props = TRUE)
   }
   res <- if (! is.null(ht1)) rbind(ht1, to_insert, copy_cell_props = copy_cell_props) else to_insert
   res <- if (! is.null(ht2)) rbind(res, ht2) else res
+
+  colspan(res)[after + 1, 1] <- colspan
 
   res
 }