+++ NEWS.md
@@ -1,8 +1,10 @@
 
 
-
 Note that huxtable attempts to follow semantic versioning (https://semver.org). Therefore, major 
 version increments reflect backwards-incompatible API changes, not necessarily big changes.
+
+# huxtable (development version)
+
 
 # huxtable 4.7.0